add azure event bus

+++ Main/SoftUnlimit.WebApi/Sources/CQRS/Event/MyEvent.cs
@@ -5,7 +5,7 @@ namespace SoftUnlimit.WebApi.Sources.CQRS.Event
 {
     public class MyEvent : VersionedEvent<Guid>
     {
-        public MyEvent(Guid id, Guid sourceId, long version, uint serviceId, string workerId, string correlationId, object command, object prevState, object currState, bool isDomainEvent, object body = null) 
+        public MyEvent(Guid id, Guid sourceId, long version, ushort serviceId, string workerId, string correlationId, object command, object prevState, object currState, bool isDomainEvent, object body = null) 
             : base(id, sourceId, version, serviceId, workerId, correlationId, command, prevState, currState, isDomainEvent, body)
         {
         }
